Migrated Global.scala to StartUpService and ErrorHandler.
