+++ app/ErrorHandler.scala
@@ -1,0 +1,25 @@
+import javax.inject.Singleton
+import play.api.http.HttpErrorHandler
+import play.api.mvc.Results._
+import play.api.mvc._
+
+import scala.concurrent._
+
+/**
+ * Manage application errors.
+ */
+@Singleton
+class ErrorHandler extends HttpErrorHandler {
+
+  def onClientError(request: RequestHeader, statusCode: Int, message: String): Future[Result] = {
+    Future.successful(
+      Status(statusCode)("A client error occurred: " + message)
+    )
+  }
+
+  def onServerError(request: RequestHeader, exception: Throwable): Future[Result] = {
+    Future.successful(
+      InternalServerError("A server error occurred: " + exception.getMessage)
+    )
+  }
+}
