+++ conf/application.conf
@@ -1,5 +1,12 @@
 # This is the main configuration file for the application.
 
+# Dependency Injection
+play.modules.enabled += "services.EagerLoaderModule"
+play.modules.enabled += "services.DI"
+
+# Support both JSON and HTML when returning errors
+play.http.errorHandler = play.api.http.HtmlOrJsonHttpErrorHandler
+
 # configuration can be specified using dot notation or object notation.
 # For example this:
 # clowder.diskStorage.path="/tmp/clowder"
